make sphere main home page
