+++ client/components/App.js
@@ -1,17 +1,17 @@
 //libraries
-import React, { Component } from 'react';
-import { Route } from 'react-router-dom';
-import { connect } from 'react-redux';
+import React, {Component} from 'react';
+import {Route} from 'react-router-dom';
+import {connect} from 'react-redux';
 import ReactDependentScript from 'react-dependent-script';
 
 //components
-import HomeSphere from './Home-Sphere';
 import Home from './Home';
+import HomeCheck from './Home-Check';
 import Nav from './Nav';
 import Map from './Map';
 
 //redux
-import { fetchNearby } from '../redux/store';
+import {fetchNearby} from '../redux/store';
 
 //styles
 import '../styles/App.css';
@@ -24,9 +24,11 @@ class App extends Component {
     });
 
     navigator.geolocation.getCurrentPosition(position => {
-      const { latitude, longitude } = position.coords;
-      this.props
-        .fetchNearby({ coordinate: { latitude, longitude }, distance: 500 }) //distance is in feet 5280ft = 1mi
+      const {latitude, longitude} = position.coords;
+      this.props.fetchNearby({
+        coordinate: {latitude, longitude},
+        distance: 500,
+      }); //distance is in feet 5280ft = 1mi
     });
   }
 
@@ -35,7 +37,7 @@ class App extends Component {
       <div>
         <Nav />
         <Route exact path="/" component={Home} />
-        <Route exact path="/homesphere" component={HomeSphere} />
+        <Route path="/homecheck" component={HomeCheck} />
         {/* <Route path="/entry" component={Entry} /> */}
         <Route
           exact
@@ -59,5 +61,5 @@ class App extends Component {
 
 export default connect(
   null,
-  { fetchNearby }
+  {fetchNearby}
 )(App);
